requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
cachetools==4.1.1         # via google-auth
certifi==2020.6.20        # via requests
cffi==1.14.3              # via google-crc32c, pynacl
chardet==3.0.4            # via requests
click==7.1.2              # via pip-tools
deprecated==1.2.10        # via pygithub
google-api-core==1.22.2   # via google-cloud-bigquery, google-cloud-core
google-auth==1.21.3       # via google-api-core
google-cloud-bigquery==1.28.0  # via -r requirements.in
google-cloud-core==1.4.1  # via google-cloud-bigquery
google-crc32c==1.0.0      # via google-resumable-media
google-resumable-media==1.0.0  # via google-cloud-bigquery
googleapis-common-protos==1.52.0  # via google-api-core
idna==2.10                # via requests
pip-tools==5.3.1          # via -r requirements.in
protobuf==3.13.0          # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.20           # via cffi
pygithub==1.55            # via -r requirements.in
pyjwt==2.4.0              # via pygithub
pynacl==1.5.0             # via pygithub
pytz==2020.1              # via google-api-core
pyyaml==5.3.1             # via -r requirements.in
requests==2.24.0          # via google-api-core, pygithub
rsa==4.6                  # via google-auth
six==1.15.0               # via google-api-core, google-auth, google-cloud-bigquery, google-resumable-media, pip-tools, protobuf
urllib3==1.25.10          # via requests
wrapt==1.12.1             # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
